semantic html and renaming components
